requirements.txt
fastapi
uvicorn
python-dotenv
twilio
requests
PyPDF2
pandas
spacy
vosk
numpy
pydub
gTTS
reportlab
faiss-cpu
SQLAlchemy
psycopg2-binary
python-multipart
Jinja2
websockets
